gemini_mcp_client/requirements.txt
google-genai>=1.0.0
mcp[cli]>=1.0.0
